TM-1181 Added logic to config to optionally delete the dot file so that we can debug
git-svn-id: https://opsvcs.tdsops.com/repos/tdstm/trunk@3061 fecba46a-3362-0410-80e5-8c76cdb905a0

+++ grails-app/controllers/TaskController.groovy
@@ -174,6 +174,7 @@ class TaskController {
 		def targetURI = grailsApplication.config.graph.targetURI
 		def dotExec = grailsApplication.config.graph.graphviz.dotCmd
 		def graphType = grailsApplication.config.graph.graphviz.graphType
+		def deleteDotFile = grailsApplication.config.graph.containsKey('deleteDotFile') ? grailsApplication.config.graph.deleteDotFile : true
 
 		def statusColor = [
 			(AssetCommentStatus.HOLD):['black', '#FFFF33'],
@@ -269,8 +270,8 @@ digraph runbook {
 	 	proc.waitFor()
 	
 		if (proc.exitValue() == 0) {
-			// Delete the dot file because we don't need it
-			dotFile.delete()
+			// Delete the dot file because we don't need it and configured to delete it automatically
+			if (deleteDotFile) dotFile.delete()
 			redirect(uri:"${targetURI}${imgFilename}")
 			
 		} else {
